chore: add schema to utils
