+++ frontend/src/components/pages/addEmployee/AddEmployeeFormProvider/AddEmployeeForm.interface.ts
@@ -15,3 +15,9 @@ export interface AddEmployeeForm {
   [addEmployeeFormNames.ladder]: Option;
   [addEmployeeFormNames.technology]: Option[];
 }
+
+export const addEmployeeFirstStepFields = [
+  addEmployeeFormNames.firstName,
+  addEmployeeFormNames.lastName,
+  addEmployeeFormNames.email,
+] as const;
