+++ frontend/src/components/pages/addEmployee/AddEmployeeFormProvider/AddEmployeeForm.tsx
@@ -5,32 +5,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 
 import { FormProvider } from '@app/components/common/FormProvider';
 import { AddEmployeeForm, addEmployeeFormNames } from './AddEmployeeForm.interface';
-import { z } from 'zod';
-
-export const addEmployeeFirstStepFields = [
-  addEmployeeFormNames.firstName,
-  addEmployeeFormNames.lastName,
-  addEmployeeFormNames.email,
-] as const;
-
-const schema = z.object({
-  [addEmployeeFormNames.firstName]: z.string().min(2, { message: 'First name must contain at least 2 characters' }),
-  [addEmployeeFormNames.lastName]: z.string().min(2, { message: 'Last name must contain at least 2 characters' }),
-  [addEmployeeFormNames.email]: z.string().regex(/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i, 'Invalid email address'),
-  [addEmployeeFormNames.ladder]: z.object({
-    name: z.string(),
-    id: z.string(),
-  }),
-  [addEmployeeFormNames.technology]: z
-    .object({
-      name: z.string(),
-      id: z.string(),
-    })
-    .array()
-    .refine((techArray) => techArray.some((tech) => tech.id && tech.id.trim() !== ''), {
-      message: 'At least one technology must be selected',
-    }),
-});
+import { addEmployeeFormSchema } from './AddEmployeeForm.utils';
 
 export const AddEmployeeFormProvider: FC<PropsWithChildren> = ({ children }) => {
   const form = useForm<AddEmployeeForm>({
@@ -42,7 +17,7 @@ export const AddEmployeeFormProvider: FC<PropsWithChildren> = ({ children }) => 
       [addEmployeeFormNames.ladder]: {},
       [addEmployeeFormNames.technology]: [],
     },
-    resolver: zodResolver(schema),
+    resolver: zodResolver(addEmployeeFormSchema),
   });
 
   return <FormProvider<AddEmployeeForm> form={form}>{children}</FormProvider>;
