+++ frontend/src/components/pages/addEmployee/AddEmployeeFormProvider/index.ts
@@ -1,3 +1,3 @@
 export { AddEmployeeFormProvider } from './AddEmployeeForm';
 export type { AddEmployeeForm } from './AddEmployeeForm.interface';
-export { addEmployeeFormNames } from './AddEmployeeForm.interface';
+export { addEmployeeFormNames, addEmployeeFirstStepFields } from './AddEmployeeForm.interface';
